Not much changed
This commit was based around finding the osx performance issue when rendering a seemingly reasonable scene. The problem occurs when the window size is increased or set to fullscreen.
CGLFlushDrawable seems to be the root of the issue. The only solution I can think of is to change the osx platform code.